+++ components/role-mgt/org.wso2.carbon.identity.role.v2.mgt.core/src/main/java/org/wso2/carbon/identity/role/v2/mgt/core/internal/RoleManagementServiceComponent.java
@@ -33,7 +33,7 @@ import org.wso2.carbon.identity.event.services.IdentityEventService;
 import org.wso2.carbon.identity.organization.management.service.OrganizationManager;
 import org.wso2.carbon.identity.role.v2.mgt.core.RoleManagementService;
 import org.wso2.carbon.identity.role.v2.mgt.core.RoleManagementServiceImpl;
-import org.wso2.carbon.identity.role.v2.mgt.core.listener.RoleManagementAuditV2Logger;
+import org.wso2.carbon.identity.role.v2.mgt.core.listener.RoleManagementV2AuditLogger;
 import org.wso2.carbon.identity.role.v2.mgt.core.listener.RoleManagementListener;
 import org.wso2.carbon.idp.mgt.IdpManager;
 import org.wso2.carbon.user.core.service.RealmService;
@@ -53,7 +53,7 @@ public class RoleManagementServiceComponent {
         try {
             BundleContext bundleContext = context.getBundleContext();
             bundleContext.registerService(RoleManagementService.class, new RoleManagementServiceImpl(), null);
-            bundleContext.registerService(RoleManagementListener.class, new RoleManagementAuditV2Logger(), null);
+            bundleContext.registerService(RoleManagementListener.class, new RoleManagementV2AuditLogger(), null);
 
             log.debug("Role V2 management service is activated.");
         } catch (Throwable e) {
@@ -173,17 +173,17 @@ public class RoleManagementServiceComponent {
 
     @Reference(
             name = "role.management.audit.v2.logger",
-            service = RoleManagementAuditV2Logger.class,
+            service = RoleManagementV2AuditLogger.class,
             cardinality = ReferenceCardinality.MULTIPLE,
             policy = ReferencePolicy.DYNAMIC,
             unbind = "unsetRoleManagementAuditV2Logger"
     )
-    protected void setRoleManagementAuditV2Logger(RoleManagementAuditV2Logger roleManagementAuditV2Logger) {
+    protected void setRoleManagementAuditV2Logger(RoleManagementV2AuditLogger roleManagementAuditV2Logger) {
 
         RoleManagementServiceComponentHolder.getInstance().addRoleManagementListener(roleManagementAuditV2Logger);
     }
 
-    protected void unsetRoleManagementAuditV2Logger(RoleManagementAuditV2Logger roleManagementListener) {
+    protected void unsetRoleManagementAuditV2Logger(RoleManagementV2AuditLogger roleManagementListener) {
 
         RoleManagementServiceComponentHolder.getInstance().setRoleManagementListenerList(null);
     }
